Generalize stack canary leaking
Make the leaking code adaptable to not only leak canaries but also
other values (e.g. return addresses) that cause an error when
modified.
Adjust the echoserver so that the leaking code does not depend on
the error message of the echoserver but on the absence of a success
message.